BandTX update

+++ STM32/Src/trx_manager.h
@@ -26,7 +26,7 @@ extern void TRX_TemporaryMute(void);
 extern void TRX_ProcessScanMode(void);
 extern void TRX_setFrequencySlowly(uint64_t target_freq);
 extern void TRX_setFrequencySlowly_Process(void);
-extern bool TRX_TX_Disabled(uint32_t freq);
+extern bool TRX_TX_Disabled(uint64_t freq);
 extern void TRX_DoFrequencyEncoder(float32_t direction, bool secondary_encoder);
 
 extern void BUTTONHANDLER_MODE_P(uint32_t parameter);
